+++ HISTORY.rst
@@ -25,6 +25,8 @@ latest (unreleased)
 
 **Bugfixes**
 
+* Fix problem with complete_name on stock locations
+* Add new "Tous les stocks" location to only get stock info from it
 * Re-use original package in return for unclaimed claims
 * Correct workflows for IN/OUT from claims (unclaimed or not)
 * Fix Swiss PP PDF label formatting